requirements.txt
beautifulsoup4==4.12.2
en_core_web_sm==3.5.0
Flask==2.3.2
locationtagger==0.0.1
nltk==3.8.1
pandas==1.5.3
python_dateutil==2.8.2
regex==2022.7.9
Requests==2.31.0
spacy==3.5.3
